split particle, system
