+++ docs/basic/system.rst
@@ -3,7 +3,7 @@
 System
 ========
 
-.. automodule:: jorbit.system
+.. automodule:: jorbit.system.system
     :members:
     :member-order: groupwise
     :undoc-members:
